Move the content of state.go to kibana_controller.go where it's used
