+++ pkg/controller/kibana/kibana_controller.go
@@ -241,3 +241,29 @@ func (r *ReconcileKibana) onDelete(obj types.NamespacedName) {
 	// Clean up watches set on secure settings
 	r.dynamicWatches.Secrets.RemoveHandlerForKey(keystore.SecureSettingsWatchName(obj))
 }
+
+// State holds the accumulated state during the reconcile loop including the response and a pointer to a Kibana
+// resource for status updates.
+type State struct {
+	Kibana  *kbv1.Kibana
+	Request reconcile.Request
+
+	originalKibana *kbv1.Kibana
+}
+
+// NewState creates a new reconcile state based on the given request and Kibana resource with the resource
+// state reset to empty.
+func NewState(request reconcile.Request, kb *kbv1.Kibana) State {
+	return State{Request: request, Kibana: kb, originalKibana: kb.DeepCopy()}
+}
+
+// UpdateKibanaState updates the Kibana status based on the given deployment.
+func (s State) UpdateKibanaState(deployment appsv1.Deployment) {
+	s.Kibana.Status.AvailableNodes = deployment.Status.AvailableReplicas
+	s.Kibana.Status.Health = kbv1.KibanaRed
+	for _, c := range deployment.Status.Conditions {
+		if c.Type == appsv1.DeploymentAvailable && c.Status == corev1.ConditionTrue {
+			s.Kibana.Status.Health = kbv1.KibanaGreen
+		}
+	}
+}
